Classes reorganization
